Add security to allow only admin users to page

+++ admin/post_comments.php
@@ -26,16 +26,30 @@
             <?php 
               if(isset($_GET['comment_post_id']))
                 {
-                  $the_post_id = $_GET['comment_post_id'];
+                  if(isset($_SESSION['user_role']))
+                    {
+                      if($_SESSION['user_role'] == "Admin")
+                        {
+                          $the_post_id = escape($_GET['comment_post_id']);
+                        }
+                      else
+                        {    
+                          header("Location: index.php");
+                        }
+                    }
+                  else
+                    {
+                         header("Location: ../index.php");
+                    }
                 }
-              $query = "SELECT * FROM comments WHERE comment_post_id = " . mysqli_real_escape_string($connection, $the_post_id). "";
+              $query = "SELECT * FROM comments WHERE comment_post_id = $the_post_id";
               $select_all_comments = mysqli_query($connection, $query);
               while($row = mysqli_fetch_assoc($select_all_comments))
                 {
                   $comment_id = $row['id'];
                   $comment_author = $row['comment_author'];
                   $comment_email = $row['comment_email'];
-                  $comment_post_id = $row['comment_post_id'];
+                  $comment_post_id = escape($row['comment_post_id']);
                   $comment_status = $row['comment_status'];
                   $comment_content = $row['comment_content'];
                   $comment_date = $row['comment_date'];
@@ -56,7 +70,7 @@
                   echo "<td>{$comment_date}</td>";
                   echo "<td><a href='post_comments.php?approve={$comment_id}&comment_post_id={$the_post_id}'>Approve</a></td>";
                   echo "<td><a href='post_comments.php?unapprove={$comment_id}&comment_post_id={$the_post_id}'>Unapprove</a></td>";
-                  echo "<td><a href='post_comments.php?delete={$comment_id}&comment_post_id={$the_post_id}'>Delete</a></td>";
+                  echo "<td><a onClick= \"javascript: return confirm('Are you sure you want to delete this comment?'); \" href='post_comments.php?delete={$comment_id}&comment_post_id={$the_post_id}'>Delete</a></td>";
                   echo "</tr>";
                 }
             ?>                  
@@ -66,29 +80,71 @@
         <?php
            if(isset($_GET['unapprove']))
              {
-               $the_comment_id = $_GET['unapprove'];
-               $query = "UPDATE comments SET comment_status = 'unapproved' WHERE id = $the_comment_id ";
-               $unapprove_comment_query = mysqli_query($connection, $query);
-               confirm_query($unapprove_comment_query);
-               header("Location: post_comments.php?comment_post_id={$the_post_id}");
+               if(isset($_SESSION['user_role']))
+                    {
+                      if($_SESSION['user_role'] == "Admin")
+                        {
+                          $the_comment_id = escape($_GET['unapprove']);
+                          $query = "UPDATE comments SET comment_status = 'unapproved' WHERE id = $the_comment_id ";
+                          $unapprove_comment_query = mysqli_query($connection, $query);
+                          confirm_query($unapprove_comment_query);
+                          header("Location: post_comments.php?comment_post_id={$the_post_id}");
+                        }
+                      else
+                        {    
+                          header("Location: index.php");
+                        }
+                    }
+                  else
+                    {
+                         header("Location: ../index.php");
+                    }
              }
 
            if(isset($_GET['approve']))
              {
-               $the_comment_id = $_GET['approve'];
-               $query = "UPDATE comments SET comment_status = 'approved' WHERE id = $the_comment_id ";
-               $approve_comment_query = mysqli_query($connection, $query);
-               confirm_query($approve_comment_query);
-               header("Location: post_comments.php?comment_post_id={$the_post_id}");
+                 if(isset($_SESSION['user_role']))
+                    {
+                      if($_SESSION['user_role'] == "Admin")
+                        {
+                          $the_comment_id = escape($_GET['approve']);
+                          $query = "UPDATE comments SET comment_status = 'approved' WHERE id = $the_comment_id ";
+                          $approve_comment_query = mysqli_query($connection, $query);
+                          confirm_query($approve_comment_query);
+                          header("Location: post_comments.php?comment_post_id={$the_post_id}");
+                        }
+                      else
+                        {    
+                          header("Location: index.php");
+                        }
+                    }
+                  else
+                    {
+                         header("Location: ../index.php");
+                    }
              }
                    
            if(isset($_GET['delete']))
              {
-               $the_comment_id = $_GET['delete'];
-               $query = "DELETE FROM comments WHERE id = $the_comment_id ";
-               $delete_comment_query = mysqli_query($connection, $query);
-               confirm_query($delete_comment_query);
-               header("Location: post_comments.php?comment_post_id={$the_post_id}");
+               if(isset($_SESSION['user_role']))
+                    {
+                      if($_SESSION['user_role'] == "Admin")
+                        {
+                          $the_comment_id = escape($_GET['delete']);
+                          $query = "DELETE FROM comments WHERE id = $the_comment_id ";
+                          $delete_comment_query = mysqli_query($connection, $query);
+                          confirm_query($delete_comment_query);
+                          header("Location: post_comments.php?comment_post_id={$the_post_id}");
+                     }
+                      else
+                        {    
+                          header("Location: index.php");
+                        }
+                    }
+                  else
+                    {
+                         header("Location: ../index.php");
+                    }
              }
         ?>
